Changed: Simplified undo commands due to simplification in `engine.model.History`. Changed: `BaseCommand#_undo` uses `engine.model.Document#transformDeltas` and uses context.
